+++ services/conversation/index.js
@@ -1,11 +1,19 @@
 const { buildConversationState, initialConversationState } = require('./conversation-state');
-const { applyDialoguePolicy, sanitizeConsultantText } = require('./dialogue-policy');
+const { applyDialoguePolicy, countQuestions, sanitizeConsultantText } = require('./dialogue-policy');
+const { determineFreedom, freedomLevel } = require('./controlled-freedom');
+const { detectNewInfo } = require('./new-info-detector');
+const { detectRepetition } = require('./repetition-detector');
 const { PERSONA } = require('./persona');
 
 module.exports = {
   PERSONA,
   applyDialoguePolicy,
   buildConversationState,
+  countQuestions,
+  detectNewInfo,
+  detectRepetition,
+  determineFreedom,
+  freedomLevel,
   initialConversationState,
   sanitizeConsultantText,
 };